[back/user] refa : router and controller update
